EWPP-667: Add entity type UI test.
